Completed working on second task

+++ painter/op.go
@@ -8,13 +8,11 @@ import (
 	"golang.org/x/image/draw"
 )
 
-// Operation змінює вхідну текстуру.
+
 type Operation interface {
-	// Do виконує зміну операції, повертаючи true, якщо текстура вважається готовою для відображення.
 	Do(t screen.Texture) (ready bool)
 }
 
-// OperationList групує список операції в одну.
 type OperationList []Operation
 
 func (ol OperationList) Do(t screen.Texture) (ready bool) {
@@ -24,14 +22,14 @@ func (ol OperationList) Do(t screen.Texture) (ready bool) {
 	return
 }
 
-// UpdateOp операція, яка не змінює текстуру, але сигналізує, що текстуру потрібно розглядати як готову.
+
 var UpdateOp = updateOp{}
 
 type updateOp struct{}
 
 func (op updateOp) Do(t screen.Texture) bool { return true }
 
-// OperationFunc використовується для перетворення функції оновлення текстури в Operation.
+
 type OperationFunc func(t screen.Texture)
 
 func (f OperationFunc) Do(t screen.Texture) bool {
@@ -39,38 +37,46 @@ func (f OperationFunc) Do(t screen.Texture) bool {
 	return false
 }
 
-// WhiteFill зафарбовує тестуру у білий колір. Може бути викоистана як Operation через OperationFunc(WhiteFill).
 func WhiteFill(t screen.Texture) {
 	t.Fill(t.Bounds(), color.White, screen.Src)
 }
 
-// GreenFill зафарбовує тестуру у зелений колір. Може бути викоистана як Operation через OperationFunc(GreenFill).
 func GreenFill(t screen.Texture) {
 	t.Fill(t.Bounds(), color.RGBA{G: 0xff, A: 0xff}, screen.Src)
 }
 
 type BgRectangle struct {
-	X1, Y1, X2, Y2 int
+	X1, Y1, X2, Y2 float64
 }
 
 func (op *BgRectangle) Do(t screen.Texture) bool {
-	t.Fill(image.Rect(op.X1, op.Y1, op.X2, op.Y2), color.RGBA{0, 255, 0, 255}, screen.Src)
+	t.Fill(image.Rect(
+		int(op.X1), int(op.Y1),
+		int(op.X2), int(op.Y2)),
+		color.RGBA{0, 255, 0, 255}, screen.Src,
+	)
 	return false
 }
 
 type Figure struct {
-	X, Y int
+	X, Y float64
 	C    color.RGBA
 }
 
 func (op *Figure) Do(t screen.Texture) bool {
-	t.Fill(image.Rect(op.X-150, op.Y-100, op.X+150, op.Y), op.C, draw.Src)
-	t.Fill(image.Rect(op.X-50, op.Y, op.X+50, op.Y+100), op.C, draw.Src)
+	t.Fill(image.Rect(
+		int(op.X-150), int(op.Y-100),
+		int(op.X+150), int(op.Y),
+	), op.C, draw.Src)
+	t.Fill(image.Rect(
+		int(op.X-50), int(op.Y),
+		int(op.X+50), int(op.Y+100),
+	), op.C, draw.Src)
 	return false
 }
 
 type Move struct {
-	X, Y    int
+	X, Y    float64
 	Figures []*Figure
 }
